Summary
This is the classic Connect 4 game we all had grown to love to play. The game is designed for two players. Each player take turn dropping one disk at a time into a six by seven slotted game board.  The first player to achieve four in a row in sequence wins.  The game is  in a draw when all 42 slots are filled. The game is then reset. Play can continue as long as it takes. In this version we are using Object Oriented Programming concepts. This is to demonstrate our knowledge and experience with Objected Oriented Programming in C#.